Fixed Double vs Float

+++ src/main/java/edu/odu/cs/cs350/TxtWriter.java
@@ -4,7 +4,7 @@ import java.util.*;
 
 public class TxtWriter {
 	
-	private Vector<Float> sizes = new Vector<Float>();
+	private Vector<Double> sizes = new Vector<Double>();
 	private Vector<String> pages = new Vector<String>();
 	
 	/**
